% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{data_download}
\alias{data_download}
\title{Download TCGA data from API}
\usage{
data_download(cohort, datatype = "RNAseq")
}
\arguments{
\item{cohort}{TCGA_CHOL}

\item{datatype}{RNA-seq}
}
\description{
Download TCGA data from API
}
\examples{
data_download("TCGA_CHOL", datatype="RNAseq")

}
