% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_gene.R
\name{plot_surv_gene}
\alias{plot_surv_gene}
\title{Plot survival curve for single gene}
\usage{
plot_surv_gene(symbol, anno = gene, count = exp, surv = survdata,
  path)
}
\arguments{
\item{symbol}{gene symbol, character}

\item{anno}{gene annotation file, output from TCGAbiolinks}

\item{count}{gene count matrix, output from TCGAbiolinks}

\item{surv}{survival data}

\item{path}{path to plot}
}
\description{
Plot survival curve for single gene
}
\examples{
chol <- data_download("TCGA-CHOL", datatype="RNA-seq")
x <- chol$x
survdata <- chol$surv
plot_surv_gene("TET2", x$genes, x$counts, survdata, "./")

}
