% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_download.R
\name{data_download}
\alias{data_download}
\title{Download TCGA data from API}
\usage{
data_download(cohort, datatype = "RNAseq")
}
\arguments{
\item{cohort}{Character, e.g. TCGA-CHOL, please refer to GDC portal for proper abbreviations.}

\item{datatype}{Character, now only support RNA-seq, e.g. RNA-seq/RNAseq}
}
\value{
List: including a DGEList object, and survival data
}
\description{
Download TCGA data from API, using TCGAbiolinks package
}
\examples{
data_download("TCGA-CHOL", datatype="RNAseq")

}
