% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_gene.R
\name{plot_surv_gene}
\alias{plot_surv_gene}
\title{Plot survival curve for single gene}
\usage{
plot_surv_gene(symbol, anno = gene, count = exp, surv = survdata,
  path)
}
\arguments{
\item{symbol}{Character, gene symbol, character}

\item{anno}{Dataframe, gene annotation file, output from TCGAbiolinks}

\item{count}{Matrix, gene count matrix, output from TCGAbiolinks}

\item{surv}{Dataframe, survival data}

\item{path}{Character, path to plot}
}
\value{
summary statistics and plot
}
\description{
Plot survival curve for single gene
}
\examples{
chol <- data_download("TCGA-CHOL", datatype="RNA-seq")
x <- chol$x
survdata <- chol$surv
plot_surv_gene("TET2", x$genes, x$counts, survdata, "./")

}
