% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv_gene.R
\name{plot_surv_gene}
\alias{plot_surv_gene}
\title{Plot survival curve for single gene}
\usage{
plot_surv_gene(symbol, anno = gene, count = exp, surv = survdata,
  path)
}
\arguments{
\item{symbol}{}

\item{anno}{}

\item{count}{}

\item{surv}{}

\item{path}{}
}
\description{
Plot survival curve for single gene
}
